refactor(weave_ts): retain Claude OTel tracer filenames
